Mejoras

+++ config/session_limits.php
@@ -2,5 +2,6 @@
 
 return [
     'idle_minutes' => (int) env('SESSION_IDLE_MINUTES', 15),
-    'max_hours' => (int) env('SESSION_MAX_HOURS', 8),
+    // Permite decimales para pruebas (ej: 0.02 ~= 72s).
+    'max_hours' => (float) env('SESSION_MAX_HOURS', 8),
 ];
